Updates to flux sequences and some minor sequencer things
- Bus T1 sequence
- Moved fluxing sequences out of the multi qubit module
- Moved distort and compensate to only one location (standard elts)
- Back to old version of minimum finding
