+++ pycqed/measurement/pulse_sequences/multi_qubit_tek_seq_elts.py
@@ -1,11 +1,12 @@
 import logging
 import numpy as np
 from copy import deepcopy
-from ..waveform_control import element
-from ..waveform_control import pulse
-from ..waveform_control import sequence
+from pycqed.measurement.waveform_control import element
+from pycqed.measurement.waveform_control import pulse
+from pycqed.measurement.waveform_control import sequence
 from pycqed.utilities.general import add_suffix_to_dict_keys
 from pycqed.measurement.pulse_sequences.standard_elements import multi_pulse_elt
+from pycqed.measurement.pulse_sequences.standard_elements import distort_and_compensate
 
 from pycqed.measurement.pulse_sequences.single_qubit_tek_seq_elts import get_pulse_dict_from_pars
 from importlib import reload
@@ -813,40 +814,3 @@ def preload_kernels_func(distortion_dict):
                     cached_kernels.update({kernel: kernel_vec})
     return output_dict
 
-
-def distort_and_compensate(element, distortion_dict):
-    """
-    Distorts an element using the contenst of a distortion dictionary.
-    The distortion dictionary should be formatted as follows.
-
-    dist_dict{'ch_list': ['chx', 'chy'],
-              'chx': np.array(.....),
-              'chy': np.array(.....)}
-    """
-
-    t_vals, outputs_dict = element.waveforms()
-    for ch in distortion_dict['ch_list']:
-        element._channels[ch]['distorted'] = True
-        length = len(outputs_dict[ch])
-        kernelvec = distortion_dict[ch]
-        outputs_dict[ch] = np.convolve(
-            outputs_dict[ch], kernelvec)[:length]
-        element.distorted_wfs[ch] = outputs_dict[ch][:len(t_vals)]
-    return element
-
-'''
-def distort_and_compensate(element, distortion_dict, preloaded_kernels):
-    t_vals, outputs_dict = element.waveforms()
-    # print(len(t_vals),t_vals[-1])
-    for ch in distortion_dict['ch_list']:
-        element._channels[ch]['distorted'] = True
-        length = len(outputs_dict[ch])
-        for kernelvec in preloaded_kernels[ch]:
-            outputs_dict[ch] = np.convolve(
-                outputs_dict[ch], kernelvec)[:length]
-
-        element.distorted_wfs[ch] = outputs_dict[ch][:len(t_vals)]
-    return element
-
-
-'''
